refactor(CO-3211): update actions type imports to use specific paths for better clarity
