+++ src/hooks/actions/use-conv-forward-as-attachment.ts
@@ -11,7 +11,8 @@ import { useTranslation } from 'react-i18next';
 import { ConversationActionsDescriptors, EditViewActions } from 'constants/index';
 import { MIMETYPE_EML } from 'helpers/attachments';
 import { useMsgForwardAsAttachmentFn } from 'hooks/actions/use-msg-forward-as-attachment';
-import { ActionFn, UIActionDescriptor, UnsavedAttachment } from 'types/index.d';
+import { ActionFn, UIActionDescriptor } from 'types/actions';
+import { UnsavedAttachment } from 'types/attachments';
 import { createEditBoard } from 'views/app/detail-panel/edit/edit-view-board';
 
 type ConvForwardAsAttachmentAction = {
