+++ src/hooks/actions/use-conv-forward.ts
@@ -9,7 +9,7 @@ import { useTranslation } from 'react-i18next';
 
 import { ConversationActionsDescriptors, EditViewActions } from 'constants/index';
 import { useMsgForwardFn } from 'hooks/actions/use-msg-forward';
-import { ActionFn, UIActionDescriptor } from 'types/index.d';
+import { ActionFn, UIActionDescriptor } from 'types/actions';
 import { createEditBoard } from 'views/app/detail-panel/edit/edit-view-board';
 
 type ConvForwardAction = {
