+++ src/hooks/actions/use-conv-set-unread.tsx
@@ -10,7 +10,7 @@ import { useTranslation } from 'react-i18next';
 import { ConversationActionsDescriptors } from 'constants/index';
 import { isDraft } from 'helpers/folders';
 import { convActionEmailStoreAction } from 'store/emails/actions/conv-action-action';
-import { ActionFn, UIActionDescriptor } from 'types/index.d';
+import { ActionFn, UIActionDescriptor } from 'types/actions';
 
 type ConvSetUnreadFunctionsParameter = {
 	ids: Array<string>;
