+++ src/hooks/actions/use-conv-show-original.tsx
@@ -9,7 +9,7 @@ import { useTranslation } from 'react-i18next';
 
 import { ConversationActionsDescriptors } from 'constants/index';
 import { isDraft, isTrash } from 'helpers/folders';
-import { ActionFn, UIActionDescriptor } from 'types/index.d';
+import { ActionFn, UIActionDescriptor } from 'types/actions';
 
 export const useConvShowOriginalFn = (firstMessageId: string, folderId: string): ActionFn => {
 	const canExecute = useCallback(
