+++ src/hooks/actions/use-msg-reply-all.tsx
@@ -10,7 +10,7 @@ import { useTranslation } from 'react-i18next';
 
 import { EditViewActions, MessageActionsDescriptors } from 'constants/index';
 import { isDraft, isSpam } from 'helpers/folders';
-import { ActionFn, UIActionDescriptor } from 'types/index.d';
+import { ActionFn, UIActionDescriptor } from 'types/actions';
 import { createEditBoard } from 'views/app/detail-panel/edit/edit-view-board';
 
 export const useMsgReplyAllFn = (messageId: string, folderId: string): ActionFn => {
