+++ src/hooks/actions/use-msg-set-not-spam.tsx
@@ -12,7 +12,7 @@ import { useNavigate } from 'react-router-dom';
 import { MAILS_ROUTE, MessageActionsDescriptors, TIMEOUTS } from 'constants/index';
 import { isSpam } from 'helpers/folders';
 import { msgActionEmailStoreAction } from 'store/emails/actions/msg-action-action';
-import { ActionFn, UIActionDescriptor } from 'types/index.d';
+import { ActionFn, UIActionDescriptor } from 'types/actions';
 
 type MsgSetNotSpam = {
 	ids: Array<string>;
