+++ src/hooks/actions/use-msg-set-spam.tsx
@@ -13,7 +13,7 @@ import { useNavigate } from 'react-router-dom';
 import { MAILS_ROUTE, MessageActionsDescriptors, TIMEOUTS } from 'constants/index';
 import { isDraft, isSpam } from 'helpers/folders';
 import { msgActionEmailStoreAction } from 'store/emails/actions/msg-action-action';
-import { ActionFn, UIActionDescriptor } from 'types/index.d';
+import { ActionFn, UIActionDescriptor } from 'types/actions';
 
 type MsgSetSpam = {
 	ids: Array<string>;
