+++ src/hooks/use-tag-dropdown-item.tsx
@@ -8,7 +8,7 @@ import React, { useMemo } from 'react';
 import { DropdownItem } from '@zextras/carbonio-design-system';
 import { includes, reduce } from 'lodash';
 
-import { UIActionAggregator } from 'types/index.d';
+import { UIActionAggregator } from 'types/actions';
 import { TagsDropdownItem } from 'ui-actions/tag-actions';
 
 export const useTagDropdownItem = (
