+++ src/views/app/folder-panel/parts/list-item-hover-actions.tsx
@@ -7,7 +7,7 @@ import React, { ReactElement, useCallback } from 'react';
 
 import { Button, Tooltip } from '@zextras/carbonio-design-system';
 
-import { UIActionDescriptor } from 'types/index.d';
+import { UIActionDescriptor } from 'types/actions';
 
 const ListItemHoverAction = ({ action }: { action: UIActionDescriptor }): ReactElement => {
 	const onClick = useCallback(
