+++ src/views/app/folder-panel/parts/multiple-selection-actions-component.tsx
@@ -15,7 +15,7 @@ import {
 } from '@zextras/carbonio-design-system';
 import { isNil, noop } from 'lodash';
 
-import { UIActionDescriptor } from 'types/index.d';
+import { UIActionDescriptor } from 'types/actions';
 
 export const MultipleSelectionActionsComponent = ({
 	actions
